+++ grapher_builder/lib/src/source/subscription.dart
@@ -1,5 +1,6 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
+import 'package:grapher_builder/src/source/class.dart';
 import 'package:grapher_builder/src/utils/annotation_data.dart';
 import 'package:grapher_builder/src/utils/buffer.dart';
 import 'package:grapher_builder/src/source/entity.dart';
@@ -13,6 +14,7 @@ class SubscriptionSource extends Entity {
   final String dartName;
   final String graphName;
 
+  late final List<MethodParameter> params;
   late final Definition result;
 
   @override
@@ -31,6 +33,7 @@ class SubscriptionSource extends Entity {
   ) {
     return [
       ...element.getters.map((e) => SubscriptionSource.parseGetter(parent, e)),
+      ...element.methods.map((e) => SubscriptionSource.parse(parent, e)),
     ].nonNulls.toList();
   }
 
@@ -51,9 +54,41 @@ class SubscriptionSource extends Entity {
       graphName: annotation.name,
     );
 
+    object.params = [];
     object.result = Definition.parse(object, result.typeArguments.first);
 
     return object;
+  }
+
+  static SubscriptionSource? parse(
+    ClassEntityType parent,
+    MethodElement element,
+  ) {
+    final annotation = SubscriptionAnnotation.peek(element);
+    if (annotation == null) {
+      return null;
+    }
+    if (!element.isStatic) {
+      throw Exception('${element.displayName} may be static!');
+    }
+    final result = element.returnType;
+    if (result is InterfaceType) {
+      final object = SubscriptionSource._(
+        resolvers: concat(parent.resolvers, annotation.resolvers),
+        parent: parent,
+        dartName: element.displayName,
+        graphName: annotation.name,
+      );
+
+      object.params = element.formalParameters
+          .map((e) => MethodParameter.parse(object, e))
+          .toList();
+
+      object.result = Definition.parse(object, result.typeArguments.first);
+
+      return object;
+    }
+    throw Exception('${element.displayName} may return Subscription<T>');
   }
 
   String generate() {
@@ -62,15 +97,30 @@ class SubscriptionSource extends Entity {
     // Function signature
     b.write('Subscription<${result.dartNameFull}> _');
     b.write([?parent?.dartName, dartName].uncapitalized);
-    b.writeln('()');
+    b.writeln('(');
+    b.block(
+      (b) =>
+          b.writeln(params.map((e) => '${e.def.dartNameFull} ${e.dartName},')),
+    );
+    b.writeln(')');
 
     b.writeln(' { ');
 
     // Body variable
     b.writeln("const body = '''");
-    b.writeln("subscription ${graphName.capitalized} {");
+    b.writeln("subscription ${graphName.capitalized}(");
+    b.block((b) => b.writeln(params.map((e) => e.queryParameter)));
+    b.writeln(') { ');
     b.block((b) {
       b.write(graphName);
+      b.write('(');
+      b.write(
+        [
+          params.firstOrNull,
+        ].nonNulls.map((e) => e.queryParameterValue).join(','),
+      );
+
+      b.write(")");
 
       final query = result.graphBody()?.split('\n');
       if (query != null) {
@@ -88,6 +138,13 @@ class SubscriptionSource extends Entity {
     // Return Query<T>
     b.writeln('return Subscription<${result.dartNameFull}>(');
     b.writeln("\tname: '$graphName',");
+    b.writeln("\tvariables: {");
+    for (final param in params) {
+      b.write("'${param.dartName}': ");
+      b.write(param.def.generateToMapValue(param.dartName));
+      b.writeln(",");
+    }
+    b.writeln("},");
     b.writeln("\tbody: body,");
 
     b.writeln("\tparserFn: (json) => ${result.generateFromMapField('json')},");